Resolução Exercícios/A2/requirements.txt
groq==1.0.0
python-dotenv==0.19.2